aider: Changed the script source path in the `index.html` file to be relative to the `index.html` file itself.
# Aider chat conversation:

USER: Change the script source path in the index.html file to be r…elative to the `index.html` file itself, not from the root of the project.

ASSISTANT: To change the script source path in the `index.html` file to be relative to the `index.html` file itself, we need to modify the `index.html` file.

Here is the change:

```html
index.html
<<<<<<< HEAD
    <script src="snake.js"></script>
=======
    <script src="./snake.js"></script>
>>>>>>> updated
```

This change makes the path to the `snake.js` file relative to the location of the `index.html` file. The `./` indicates that the `snake.js` file is in the same directory as the `index.html` file.